Se guardan los cambios seleccionados en los dropdown al cerrar y volver a abrir el modal
